FEAT - API integration done for GET and POST transactions

+++ server/src/TransactionService/DTOs/AddTransactionDto.cs
@@ -3,8 +3,8 @@ namespace TransactionService.DTOs;
 public class AddTransactionDto
 {
     public string CurrencyName { get; set; } = string.Empty;
-    public int Quantity { get; set; }
-    public int Price { get; set; }
+    public double Quantity { get; set; }
+    public double Price { get; set; }
     public bool IsBuy { get; set; }
     public Guid UserId { get; set; }
 }
